+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [3.4.0-beta1] - tba
-- [new] #1305 - Added participation indicator in effective shares list (#2037) 
+- [new] #1305 - Added participation indicator in effective shares list (#2037)
+- [change] #2038 - validate token in router and reroute before entering public page
 
 ## [3.3.0] - 2021-10-10
 - Added email addresses to external shares in the shares tab for the owner